* Fixed runes not performing properly (Lyfus Uppus rank 3, Magikus Uppus rank 3, and Zerko Killus below rank 3)

+++ source/mystic/runes.cpp
@@ -220,7 +220,7 @@ const char* RuneName(Rune r)
 		return "????";
 }
 
-int RuneValue(Rune r)
+float RuneValue(Rune r)
 {
 	if (!RuneEquipped(r))
 		return 0;
